+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @embeddable/remarkable-ui
 
+## 0.1.36
+
+### Patch Changes
+
+- 424886e: Enable Grouping on Select Fields
+
+## 0.1.35
+
+### Patch Changes
+
+- 18fb3bd: Fix an issue with MultiSelect No Options message
+
 ## 0.1.34
 
 ### Patch Changes